refactor(desktop): extract archive lifecycle helpers
Problem:
- The archive reliability fix added lines to several ratcheted legacy files, causing repolint to reject the PR.

Root cause:
- Runtime teardown, archive transaction, animation, state, and regression helpers were implemented inline in files already above the repository size budgets.

Fix:
- Move each new owner-level concern into a focused Go or TypeScript module.
- Split archive transaction phases so every function remains within the function-size budget.
- Preserve the existing lock order, cleanup markers, generation checks, and UI behavior without widening lint baselines.

Verification:
- go run ./tools/repolint
- go test ./internal/control -count=1
- cd desktop && go test ./... -count=1
- cd desktop && go test -race . -run TrashTopic\|WaitDestroyHandles\|FriendlySessionFileError -count=1
- cd desktop/frontend && pnpm build
- Focused frontend tests and local browser click verification
