+++ desktop/frontend/src/components/ApprovalModal.tsx
@@ -17,49 +17,7 @@ import {
   pickInlineFileReference,
   useFileReferenceMenu,
 } from "./FileReferenceMenu";
-
-export type ShelfExitEase = "power2.in" | "power2.out";
-
-export function shelfExitEasing(ease: ShelfExitEase): string {
-  return ease === "power2.in"
-    ? "cubic-bezier(0.55, 0.085, 0.68, 0.53)"
-    : "cubic-bezier(0.2, 0.72, 0.2, 1)";
-}
-
-export function animateShelfExit(
-  el: HTMLDivElement,
-  options: { opacity: number; y: number; duration: number; ease: ShelfExitEase; onComplete: () => void },
-) {
-  let completed = false;
-  const complete = () => {
-    if (completed) return;
-    completed = true;
-    options.onComplete();
-  };
-  if (typeof el.animate !== "function") {
-    complete();
-    return;
-  }
-  try {
-    const animation = el.animate(
-      [
-        { opacity: 1, transform: "translateY(0)" },
-        { opacity: options.opacity, transform: `translateY(${options.y}px)` },
-      ],
-      {
-        duration: options.duration * 1000,
-        easing: shelfExitEasing(options.ease),
-      },
-    );
-    animation.onfinish = complete;
-    animation.oncancel = complete;
-    void animation.finished.then(complete, complete);
-  } catch {
-    // Animation support is cosmetic. A rejected/partial Web Animations
-    // implementation must never swallow the approval action itself.
-    complete();
-  }
-}
+import { animateShelfExit } from "../lib/approvalAnimation";
 
 function requiresFreshHumanApproval(tool: string): boolean {
   return tool === "remember" || tool === "forget" || tool === "exit_plan_mode" || tool === "sandbox_escape" || tool === "config_write";
