+++ desktop/tabs.go
@@ -7660,242 +7660,6 @@ func (a *App) SetTopicPinned(topicID string, pinned bool) error {
 	return nil
 }
 
-var (
-	errTopicHasActiveWork = errors.New("wait for the session to finish, answer pending prompts, and stop background jobs before archiving this topic")
-	errTopicArchiveBusy   = errors.New("Reasonix is finishing another session change — wait a moment and retry archiving")
-)
-
-// TrashTopic removes an idle topic from the project tree and moves its saved
-// session records into the session trash. Idle in-process runtimes are detached
-// first, so their autosave cannot recreate state after the topic is gone.
-func (a *App) TrashTopic(topicID string) error {
-	return friendlySessionFileError(a.trashTopic(topicID))
-}
-
-func (a *App) topicHasActiveRuntimeWork(topicID string) bool {
-	a.mu.RLock()
-	defer a.mu.RUnlock()
-	for _, tabs := range []map[string]*WorkspaceTab{a.tabs, a.detachedSessions} {
-		for _, tab := range tabs {
-			if tab != nil && tab.TopicID == topicID && tab.hasActiveRuntimeWork() {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func (a *App) trashTopic(topicID string) (retErr error) {
-	if strings.TrimSpace(topicID) == "" {
-		return fmt.Errorf("topicID is required")
-	}
-
-	started := time.Now()
-	phase := "start"
-	targetCount := 0
-	runtimeCount := 0
-	defer func() {
-		outcome := "ok"
-		if retErr != nil {
-			outcome = "failed"
-		}
-		slog.Debug("desktop: topic archive timing",
-			"outcome", outcome,
-			"phase", phase,
-			"total_ms", time.Since(started).Milliseconds(),
-			"target_count", targetCount,
-			"runtime_count", runtimeCount,
-		)
-	}()
-
-	var fallback fallbackRuntimeTarget
-	var changedDirs []string
-	if err := func() error {
-		phase = "runtime_lock"
-		releaseRuntime, ok := a.tryLockRuntimeMutation("trash-topic")
-		if !ok {
-			return errTopicArchiveBusy
-		}
-		defer releaseRuntime()
-		phase = "removal_lock"
-		if !a.sessionRemovalMu.TryLock() {
-			return errTopicArchiveBusy
-		}
-		defer a.sessionRemovalMu.Unlock()
-		phase = "active_work_check"
-		if a.topicHasActiveRuntimeWork(topicID) {
-			return errTopicHasActiveWork
-		}
-
-		phase = "target_scan"
-		targets, err := a.topicTrashTargets(topicID)
-		if err != nil {
-			return err
-		}
-		targetCount = len(targets)
-		for _, target := range targets {
-			changedDirs = append(changedDirs, target.dir)
-		}
-
-		phase = "snapshot"
-		removed := a.captureTopicRuntimeBindings(topicID)
-		runtimeCount = len(removed)
-		if err := a.snapshotTopicRuntimeBindings(removed); err != nil {
-			return err
-		}
-
-		// Publish durable logical-removal markers before detaching runtimes or
-		// moving the first artifact. A crash or partial filesystem failure can
-		// then converge through the existing startup reconciler without losing
-		// the conversation. The legacy "delete" operation value is retained so
-		// supported older binaries continue to move these sessions to trash.
-		phase = "mark_cleanup_pending"
-		marked := make([]string, 0, len(targets))
-		rollbackMarkers := func() {
-			for _, path := range marked {
-				if err := agent.ClearCleanupPending(path); err != nil {
-					slog.Warn("desktop: rollback topic archive marker failed", "err", err)
-				}
-			}
-		}
-		for _, target := range targets {
-			if err := agent.MarkCleanupPending(target.sessionPath, "delete"); err != nil {
-				rollbackMarkers()
-				return err
-			}
-			marked = append(marked, target.sessionPath)
-		}
-
-		phase = "detach_runtimes"
-		nextFallback, unchanged := a.removeTopicRuntimeBindingsIfUnchanged(topicID, removed)
-		if !unchanged {
-			rollbackMarkers()
-			return errTopicArchiveBusy
-		}
-		fallback = nextFallback
-		a.finalizeRemovedTopicRuntimes(removed)
-		destroyBegun := false
-		closedRemoved := map[control.SessionAPI]bool{}
-		defer func() {
-			if destroyBegun {
-				a.closeRemainingRemovedSessionRuntimesAfterDestroyAdmissionHeld(removed, closedRemoved)
-				return
-			}
-			a.closeRemainingRemovedSessionRuntimesAdmissionHeld(removed, closedRemoved)
-		}()
-
-		for _, target := range targets {
-			phase = "teardown"
-			destroys := a.destroyHandlesForSession(target.dir, target.sessionPath, removed)
-			if len(destroys) > 0 {
-				destroyBegun = true
-			}
-			teardownTimedOut := waitDestroyHandles(destroys)
-			a.closeRemovedSessionRuntimesForSessionAfterDestroyAdmissionHeld(removed, target.dir, target.sessionPath, closedRemoved)
-			if teardownTimedOut {
-				go delayedDesktopSessionTrash(target.dir, target.sessionPath, target.key, destroys)
-			} else {
-				phase = "move_artifacts"
-				err := trashSessionArtifacts(target.dir, target.sessionPath, target.key)
-				finishDestroyHandles(destroys)
-				if err != nil {
-					return err
-				}
-			}
-		}
-		phase = "delete_topic_metadata"
-		return a.deleteTopic(topicID)
-	}(); err != nil {
-		return err
-	}
-	if fallback.needs {
-		phase = "open_fallback"
-		fallback.topicID = ""
-		if err := a.openFallbackRuntime(fallback); err != nil {
-			// The archive transaction is already committed. A blank-surface boot
-			// failure must not report the successful archive as failed or invite a
-			// duplicate destructive retry.
-			slog.Warn("desktop: open fallback after topic archive failed", "err", err)
-		}
-	}
-	phase = "notify"
-	if len(changedDirs) > 0 {
-		a.emitProjectTreeChangedForSessionDirs(changedDirs...)
-	} else {
-		a.emitProjectTreeMetadataChanged()
-	}
-	return nil
-}
-
-type topicTrashTarget struct {
-	dir         string
-	sessionPath string
-	key         string
-}
-
-func (a *App) topicTrashTargets(topicID string) ([]topicTrashTarget, error) {
-	topicID = strings.TrimSpace(topicID)
-	var targets []topicTrashTarget
-	seen := map[string]bool{}
-	addTarget := func(dir, path string) error {
-		sessionPath, key, err := validateSessionPath(dir, path)
-		if err != nil {
-			return err
-		}
-		id := dir + "\x00" + sessionPath
-		if seen[id] {
-			return nil
-		}
-		seen[id] = true
-		if err := validateSessionTrashTarget(dir, sessionPath, key); err != nil {
-			return err
-		}
-		targets = append(targets, topicTrashTarget{dir: dir, sessionPath: sessionPath, key: key})
-		return nil
-	}
-	for _, dir := range a.knownSessionDirs() {
-		index, err := topicSessionIndexForDir(dir)
-		if err != nil {
-			return nil, err
-		}
-		for _, match := range index.byTopic[topicID] {
-			if agent.IsCleanupPending(match.path) {
-				continue
-			}
-			if err := addTarget(dir, match.path); err != nil {
-				return nil, err
-			}
-		}
-	}
-	a.mu.RLock()
-	var runtimeTargets []struct {
-		dir  string
-		path string
-	}
-	for _, tab := range a.runtimeTabsLocked() {
-		if tab == nil || tab.TopicID != topicID {
-			continue
-		}
-		if path := canonicalTabSessionPath(tab.currentSessionPath()); path != "" {
-			dir := tabSessionDir(tab)
-			if filepath.IsAbs(path) {
-				dir = filepath.Dir(path)
-			}
-			runtimeTargets = append(runtimeTargets, struct {
-				dir  string
-				path string
-			}{dir: dir, path: path})
-		}
-	}
-	a.mu.RUnlock()
-	for _, target := range runtimeTargets {
-		if err := addTarget(target.dir, target.path); err != nil {
-			return nil, err
-		}
-	}
-	return targets, nil
-}
-
 // ListProjectTree builds the sidebar tree: project folders each containing
 // their topics, plus a Global section.
 // topicSummary is used by ListProjectTree and mergeSessionInfos to track
